+++ ydb/library/benchmarks/queries/tpcds/pg/q54.sql
@@ -45,7 +45,7 @@ with my_customers as (
  group by c_customer_sk
  )
  , segments as
- (select cast((revenue/50) as int) as segment
+ (select cast((revenue/50::numeric) as int) as segment
   from   my_revenue
  )
   select  segment, count(*) as num_customers, segment*50 as segment_base
