+++ ydb/library/benchmarks/queries/tpcds/pg/q65.sql
@@ -21,7 +21,7 @@ select
  	where ss_sold_date_sk = d_date_sk and d_month_seq between 1186 and 1186+11
  	group by ss_store_sk, ss_item_sk) sc
  where sb.ss_store_sk = sc.ss_store_sk and
-       sc.revenue <= 0.1 * sb.ave and
+       sc.revenue <= 0.1::numeric * sb.ave and
        s_store_sk = sc.ss_store_sk and
        i_item_sk = sc.ss_item_sk
  order by s_store_name, i_item_desc
