+++ ydb/library/benchmarks/queries/tpcds/pg/q91.sql
@@ -25,7 +25,7 @@ and     d_moy                   = 12
 and     ( (cd_marital_status       = 'M' and cd_education_status     = 'Unknown')
         or(cd_marital_status       = 'W' and cd_education_status     = 'Advanced Degree'))
 and     hd_buy_potential like 'Unknown%'
-and     ca_gmt_offset           = -7
+and     ca_gmt_offset           = -7::numeric
 group by cc_call_center_id,cc_name,cc_manager,cd_marital_status,cd_education_status
 order by sum(cr_net_loss) desc;
